requirements.txt
fastapi==0.95.1
torch==2.0.0
Pillow==9.5.0
numpy==1.24.3
uvicorn==0.22.0
opencv-python-headless==4.7.0.72
python-multipart==0.0.6
pandas==2.0.1